cli: rename `--destination` to `--onto`
As we've discussed a few times, `--onto` seems like a clearer name for
what we currently call `--destination`, especially since
`--insert-before` and `--insert-after` are other kinds of
destinations. It's telling that `RebaseDestinationArgs` already
exists for grouping the `-d/-A/-B` arguments.

This patch adds replaces `--destination/-d` by `--onto/-o` as the
canonical form, while keeping the former as a visible alias for now. I
expect this transition will take a long time, especially for Mercurial
converts, so I think it will be useful to keep the old name visible
for a while so users who look for it can find it and realize that it
has a new name. We could add deprecation warnings but I'm leaning
towards giving users some months to slowly update their muscle memory
before we start warning about it.

This rename makes `jj rebase --source` less symmetric with `jj rebase
--destination`. However, since there is also `--branch` and `-r` for
selecting the source commits, I think `--source` should already not be
interpreted as the counterpart to "destination". I personally think we
should rename that argument too, but I don't think it needs to be done
at the same time, though it would be good to do it soon if we can find
a good name for it.